pyproject.toml
[project]
name = "notification-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "redis>=7.3.0",
    "resend>=2.24.0",
    "strawberry-graphql>=0.311.1",
    "uvicorn>=0.42.0",
]
